+++ pointcloudset/__init__.py
@@ -1,4 +1,11 @@
 from .dataset import Dataset
+from .io.dataset.topics import list_pointcloud_topics
 from .pointcloud import PointCloud
 
 __version__ = "0.10.0"
+
+__all__ = [
+    "Dataset",
+    "PointCloud",
+    "list_pointcloud_topics",
+]